refactor(avm)!: migrate aztec_hash_policy to aztec constants
