+++ barretenberg/cpp/src/CMakeLists.txt
@@ -122,6 +122,7 @@ endif()
 if(NOT FUZZING AND NOT WASM AND NOT BB_LITE)
     # Fuzzing preset cannot be built with world_state as world_state cannot compile with MULTITHREADING=OFF
     # Mobile builds exclude these modules that require LMDB or aren't needed on mobile
+    add_subdirectory(barretenberg/aztec)
     add_subdirectory(barretenberg/world_state)
     # NOTE: Do not conditionally base this on the AVM flag as it defines a necessary vm2_sim library.
     add_subdirectory(barretenberg/vm2)
@@ -132,6 +133,7 @@ endif()
 if(FUZZING_AVM)
     if(FUZZING)
         # Only add these if they weren't added above (when NOT FUZZING AND NOT WASM)
+        add_subdirectory(barretenberg/aztec)
         add_subdirectory(barretenberg/world_state)
         add_subdirectory(barretenberg/vm2)
     endif()
@@ -205,6 +207,7 @@ if(NOT WASM AND NOT FUZZING AND NOT BB_LITE)
     list(APPEND BARRETENBERG_TARGET_OBJECTS $<TARGET_OBJECTS:crypto_merkle_tree_objects>)
     list(APPEND BARRETENBERG_TARGET_OBJECTS $<TARGET_OBJECTS:lmdblib_objects>)
     list(APPEND BARRETENBERG_TARGET_OBJECTS $<TARGET_OBJECTS:api_objects>)
+    list(APPEND BARRETENBERG_TARGET_OBJECTS $<TARGET_OBJECTS:aztec_objects>)
     list(APPEND BARRETENBERG_TARGET_OBJECTS $<TARGET_OBJECTS:world_state_objects>)
 endif()
