+++ barretenberg/cpp/src/barretenberg/vm2/constraining/relations/indexed_tree_check.test.cpp
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-#include "barretenberg/crypto/merkle_tree/aztec_hash_policy.hpp"
+#include "barretenberg/aztec/aztec_hash_policy.hpp"
 #include "barretenberg/crypto/poseidon2/poseidon2.hpp"
 #include "barretenberg/vm2/common/aztec_constants.hpp"
 #include "barretenberg/vm2/constraining/flavor_settings.hpp"
@@ -175,7 +175,7 @@ TEST(IndexedTreeCheckConstrainingTest, PositiveWriteAppend)
 
     FF value = 100;
     FF low_value = 40;
-    TestMemoryTree<crypto::merkle_tree::AztecMerkleHashPolicy> tree(8, NULLIFIER_TREE_HEIGHT);
+    TestMemoryTree<aztec::AztecMerkleHashPolicy> tree(8, NULLIFIER_TREE_HEIGHT);
 
     IndexedTreeLeafData low_leaf = {
         .leaf_separator = FF(DOM_SEP__NULLIFIER_LEAF), .value = low_value, .next_value = value + 1, .next_index = 10
