+++ barretenberg/cpp/src/barretenberg/vm2/simulation/lib/written_slots_tree.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "barretenberg/crypto/merkle_tree/aztec_hash_policy.hpp"
+#include "barretenberg/aztec/aztec_hash_policy.hpp"
 #include "barretenberg/vm2/common/field.hpp"
 #include "barretenberg/vm2/simulation/interfaces/db.hpp"
 #include "barretenberg/vm2/simulation/lib/indexed_memory_tree.hpp"
@@ -36,8 +36,7 @@ struct WrittenPublicDataSlotLeafValue {
     static std::string name();
 };
 
-using WrittenPublicDataSlotsTree =
-    IndexedMemoryTree<WrittenPublicDataSlotLeafValue, crypto::merkle_tree::AztecMerkleHashPolicy>;
+using WrittenPublicDataSlotsTree = IndexedMemoryTree<WrittenPublicDataSlotLeafValue, aztec::AztecMerkleHashPolicy>;
 
 using WrittenPublicDataSlotsTreeLeafPreimage = IndexedLeaf<WrittenPublicDataSlotLeafValue>;
