+++ barretenberg/cpp/src/barretenberg/world_state/fork.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include "barretenberg/aztec/aztec_hash_policy.hpp"
 #include "barretenberg/crypto/merkle_tree/append_only_tree/content_addressed_append_only_tree.hpp"
-#include "barretenberg/crypto/merkle_tree/aztec_hash_policy.hpp"
 #include "barretenberg/crypto/merkle_tree/indexed_tree/content_addressed_indexed_tree.hpp"
 #include "barretenberg/crypto/merkle_tree/node_store/cached_content_addressed_tree_store.hpp"
 #include "barretenberg/world_state/tree_with_store.hpp"
@@ -11,7 +11,7 @@
 
 namespace bb::world_state {
 
-using HashPolicy = crypto::merkle_tree::AztecMerkleHashPolicy;
+using HashPolicy = aztec::AztecMerkleHashPolicy;
 
 using FrStore = crypto::merkle_tree::ContentAddressedCachedTreeStore<fr>;
 using FrTree = crypto::merkle_tree::ContentAddressedAppendOnlyTree<FrStore, HashPolicy>;
